Update armTemplateWVD.json
Added the exit codes 0 and 267014 as valid exit codes to the FSlogix install script block.
"validExitCodes": [0,267014]

During the AVD build, the install of FSLOGIX can exit with the code 267014. This is a valid and successful exit code however because its not a common or known sucessful exit code, the build process can fail.